pyproject.toml
[tool.poetry]
name = "ai-devs-3"
version = "0.1.0"
description = ""
authors = ["Rafal Jamro <rafal.jamro@sirocco.eu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.9.2"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
